p3_mds::util

Function apply_circulant

pub fn apply_circulant<R: PrimeCharacteristicRing, const N: usize>(
    circ_matrix: &[u64; N],
    input: &[R; N],
) -> [R; N]

Given the first row circ_matrix of an NxN circulant matrix, say C, return the product C*input.

NB: This function is a naive implementation of the n² evaluation. It is a placeholder until we have FFT implementations for all combinations of field and size.